Move 'OverlaySchedule' into its own module.
'LedgerState' is already a giant grab-bag of functionality. This is a
good excuse to extract some of it.
